requirements.txt
matplotlib
pocket-tts
requests
python-osc
python-dotenv